Expressions lecture complete

+++ module-1/03_Expressions/student-lecture/Lecture/04_ReturnDoubleOfTwo.cs
@@ -12,8 +12,8 @@ namespace Lecture
          4. This method currently returns an int. Change it so that it returns a double.
             TOPIC: Return Types
          */
-        public int ReturnDoubleOfTwo()
-        {
+        public double ReturnDoubleOfTwo()
+            {
             return 2.0;
         }
     }
